Provide results in tiles to reduce requests
- Use tile-like approach; return a list of results for an area that contain the requested point.
- Define basic specification for new resource.

Other Notes:

The URL is /results/:algorithm-name/tile?x=:number&y=:number is similar to the way lcmap-landsat
provides data. I like that the algorithm name is part of the URL because it basically means,
"this algorithm is a network resource." By adding a "tile" subordinate resource we can provide
different chunks of data. Later, we can add another resource for retrieving arbitrary areas...